# apps/server/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OTL Server",
    "version": "1.0.0",
    "description": "OTL PLUS main Server의 Swagger 문서입니다."
  },
  "paths": {
...
    "/session/login": {
      "get": {
        "summary": "AuthController.user_login",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "null"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "name": "next",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "social_login",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/session/login/callback": {
      "get": {
        "summary": "AuthController.loginCallback",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "null"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/session/refresh": {
      "post": {
        "summary": "AuthController.refreshToken",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IUser.TokenResponse"
                }
              }
            }
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IUser.TokenDto"
              }
            }
          }
        }
      }
    },
...
    "/session/logout": {
      "get": {
        "summary": "AuthController.logout",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "null"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "name": "next",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/sync/scholarDB": {
      "post": {
        "summary": "SyncController.syncScholarDB",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ISync.ScholarDBBody"
              }
            }
          }
        }
      }
    },
    "/api/sync/examtime": {
      "post": {
        "summary": "SyncController.syncExamtime",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ISync.ExamtimeBody"
              }
            }
          }
        }
      }
    },
    "/api/sync/classtime": {
      "post": {
        "summary": "SyncController.syncClasstime",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ISync.ClasstimeBody"
              }
            }
          }
        }
      }
    },
    "/api/sync/takenLecture": {
      "post": {
        "summary": "SyncController.syncTakenLecture",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ISync.TakenLectureBody"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ISync.ClasstimeBody": {
        "properties": {
          "year": {
            "type": "integer"
          },
          "semester": {
            "type": "number",
            "enum": [1, 2, 3, 4]
          },
          "classtimes": {
            "type": "array"
          }
        },
        "type": "object",
        "required": ["year", "semester", "classtimes"]
      },
      "ISync.TakenLectureBody": {
        "properties": {
          "year": {
            "type": "integer"
          },
          "semester": {
            "type": "number",
            "enum": [1, 2, 3, 4]
          },
          "attend": {
            "type": "array"
          }
        },
        "type": "object",
        "required": ["year", "semester", "attend"]
      },
      "ISync.ExamtimeBody": {
        "properties": {
          "year": {
            "type": "integer"
          },
          "semester": {
            "type": "number",
            "enum": [1, 2, 3, 4]
          },
          "examtimes": {
            "type": "array"
          }
        },
        "type": "object",
        "required": ["year", "semester", "examtimes"]
      },
      "ISync.ScholarDBBody": {
        "properties": {
          "year": {
            "type": "integer"
          },
          "semester": {
            "type": "number",
            "enum": [1, 2, 3, 4]
          },
          "lectures": {},
          "charges": {}
        },
        "type": "object",
        "required": ["year", "semester", "lectures", "charges"]
      },
...
      "IUser.TokenDto": {
        "properties": {
          "token": {
            "minLength": 1,
            "type": "string"
          }
        },
        "type": "object",
        "required": ["token"]
      },
      "IUser.TokenResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        },
        "required": ["accessToken", "refreshToken"],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "jwtCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken"
      },
      "sidCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "AUTH-SID"
      },
      "jwtHeaderAccessToken": {
        "type": "http",
        "scheme": "bearer",
        "name": "accessToken"
      },
      "jwtHeaderRefreshToken": {
        "type": "http",
        "scheme": "bearer",
        "name": "X-REFRESH-TOKEN"
      },
      "sidHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-AUTH-SID"
      }
    }
  },
  "security": [
    {
      "jwtCookie": [],
      "sidCookie": [],
      "jwtHeaderAccessToken": [],
      "jwtHeaderRefreshToken": [],
      "sidHeader": []
    }
  ]
}